Restructure the REST client add pagination detector
