+++ sources/rest_api/__init__.py
@@ -1,8 +1,11 @@
 """Generic API Source"""
+from typing import TypedDict, Optional, Dict, List, Any
 
 import dlt
 
-from api_client import APIClient, JSONResponsePaginator, HeaderLinkPaginator, BearerTokenAuth
+from .client import RESTClient
+from .paginators import JSONResponsePaginator, HeaderLinkPaginator
+from .auth import BearerTokenAuth
 
 
 PAGINATOR_MAP = {
@@ -11,8 +14,46 @@ PAGINATOR_MAP = {
 }
 
 
+class AuthConfig(TypedDict, total=False):
+    token: str
+
+
+class ClientConfig(TypedDict, total=False):
+    base_url: str
+    auth: Optional[AuthConfig]
+    default_paginator: Optional[str]
+
+
+class ResourceConfig(TypedDict, total=False):
+    primary_key: str
+    write_disposition: str
+
+
+class IncrementalConfig(TypedDict, total=False):
+    cursor_path: str
+    initial_value: str
+    param: str
+
+
+class EndpointConfig(TypedDict):
+    params: Dict[str, Any]
+    resource: ResourceConfig
+    incremental: Optional[IncrementalConfig]
+    method: str
+    paginator: str
+
+
+class RESTAPIConfig(TypedDict):
+    client: ClientConfig
+    endpoints: Dict[str, EndpointConfig]
+
+
 def create_paginator(paginator_config):
     return PAGINATOR_MAP.get(paginator_config, lambda: None)()
+
+
+def create_auth(auth_config):
+    return BearerTokenAuth(auth_config.get("token")) if auth_config else None
 
 
 def make_client_config(config):
@@ -24,18 +65,29 @@ def make_client_config(config):
     }
 
 
-def setup_incremental_object(config):
+def setup_incremental_object(request_params, incremental_config):
+    for key, value in request_params.items():
+        if isinstance(value, dlt.sources.incremental):
+            return value, key
+
+    return setup_incremental_object_from_config(incremental_config)
+
+
+def setup_incremental_object_from_config(config):
     return (
-        dlt.sources.incremental(
-            config.get("cursor_path"),
-            initial_value=config.get("initial_value")
-        ),
-        config.get("param")
-    ) if config else (None, None)
+        (
+            dlt.sources.incremental(
+                config.get("cursor_path"), initial_value=config.get("initial_value")
+            ),
+            config.get("param"),
+        )
+        if config
+        else (None, None)
+    )
 
 
 @dlt.source
-def rest_api_source(config):
+def rest_api_source(config: RESTAPIConfig):
     """
     Creates and configures a REST API source for data extraction.
 
@@ -58,16 +110,19 @@ def rest_api_source(config):
         })
     """
 
-    client = APIClient(**make_client_config(config))
+    client = RESTClient(**make_client_config(config))
 
     for endpoint, endpoint_config in config["endpoints"].items():
         request_params = endpoint_config.get("params", {})
         resource_config = endpoint_config.get("resource", {})
 
-        incremental_object, incremental_param = setup_incremental_object(endpoint_config.get("incremental"))
+        incremental_object, incremental_param = setup_incremental_object(
+            request_params, endpoint_config.get("incremental")
+        )
 
-        def paginate_resource(method, path, params, paginator,
-                            incremental_object=incremental_object):
+        def paginate_resource(
+            method, path, params, paginator, incremental_object=incremental_object
+        ):
             if incremental_object:
                 params[incremental_param] = incremental_object.last_value
