+++ sources/rest_api/auth.py
@@ -1,0 +1,9 @@
+from requests.auth import AuthBase
+
+class BearerTokenAuth(AuthBase):
+    def __init__(self, token: str):
+        self.token = token
+
+    def __call__(self, request):
+        request.headers["Authorization"] = f"Bearer {self.token}"
+        return request
