+++ sources/rest_api/utils.py
@@ -1,0 +1,4 @@
+def join_url(base_url: str, path: str) -> str:
+    if not base_url.endswith("/"):
+        base_url += "/"
+    return base_url + path.lstrip("/")
