+++ sources/rest_api_pipeline.py
@@ -6,13 +6,16 @@ from rest_api import rest_api_source
 # dlt Requests:
 #
 
+
 @dlt.resource(
     table_name="issues",
     write_disposition="merge",
     primary_key="id",
 )
 def get_issues(
-    updated_at = dlt.sources.incremental("updated_at", initial_value="1970-01-01T00:00:00Z")
+    updated_at=dlt.sources.incremental(
+        "updated_at", initial_value="1970-01-01T00:00:00Z"
+    )
 ):
     url = (
         f"https://api.github.com/repos/dlt-hub/dlt/issues"
@@ -30,55 +33,62 @@ def get_issues(
             break
         url = response.links["next"]["url"]
 
+
 #
 # REST Source:
 #
 
+
 def load_github():
     pipeline = dlt.pipeline(
-        pipeline_name="rest_api_github", destination="duckdb", dataset_name="rest_api_data"
+        pipeline_name="rest_api_github",
+        destination="duckdb",
+        dataset_name="rest_api_data",
     )
 
-    github_source = rest_api_source({
-        "client": {
-            "base_url": "https://api.github.com/repos/dlt-hub/dlt/",
-            "default_paginator": "header_links",
-        },
-        "endpoints": {
-            "issues/comments": {
-                "params": {
-                    "per_page": 100,
+    github_source = rest_api_source(
+        {
+            "client": {
+                "base_url": "https://api.github.com/repos/dlt-hub/dlt/",
+                "default_paginator": "header_links",
+                # "auth": {
+                #     "token": dlt.secrets['token'],
+                # }
+            },
+            "endpoints": {
+                "issues/comments": {
+                    "params": {
+                        "per_page": 100,
+                        "since": dlt.sources.incremental(
+                            "updated_at", initial_value="2024-01-25T11:21:28Z"
+                        ),
+                    },
+                    "resource": {
+                        "primary_key": "id",
+                        "write_disposition": "merge",
+                    },
                 },
-                "resource": {
-                    "primary_key": "id",
-                    "write_disposition": "merge",
-                },
-                "incremental": {
-                    "cursor_path": "updated_at",
-                    "initial_value": "2024-01-25T11:21:28Z",
-                    "param": "since",
+                "issues": {
+                    "params": {
+                        "per_page": 100,
+                        "sort": "updated",
+                        "direction": "desc",
+                        "state": "open",
+                    },
+                    "resource": {
+                        "primary_key": "id",
+                        "write_disposition": "merge",
+                    },
+                    "incremental": {
+                        "cursor_path": "updated_at",
+                        "initial_value": "2024-01-25T11:21:28Z",
+                        "param": "since",
+                        # also, todo: "transform": to_iso8601,
+                    },
                 },
             },
-            "issues": {
-                "params": {
-                    "per_page": 100,
-                    "sort": "updated",
-                    "direction": "desc",
-                    "state": "open",
-                },
-                "resource": {
-                    "primary_key": "id",
-                    "write_disposition": "merge",
-                },
-                "incremental": {
-                    "cursor_path": "updated_at",
-                    "initial_value": "2024-01-25T11:21:28Z",
-                    "param": "since",
-                    # also, todo: "transform": to_iso8601,
-                },
-            },
-        },
-    })
+        }
+    )
 
     load_info = pipeline.run(github_source)
     print(load_info)
@@ -86,32 +96,36 @@ def load_github():
 
 def load_pokemon():
     pipeline = dlt.pipeline(
-        pipeline_name="rest_api_pokemon", destination="duckdb", dataset_name="rest_api_data"
+        pipeline_name="rest_api_pokemon",
+        destination="duckdb",
+        dataset_name="rest_api_data",
     )
 
-    pokemon_source = rest_api_source({
-        "client": {
-            "base_url": "https://pokeapi.co/api/v2/",
-            "default_paginator": "json_links",
-        },
-        "endpoints": {
-            "pokemon": {
-                "params": {
-                    "limit": 1000, # Default page size is 20
+    pokemon_source = rest_api_source(
+        {
+            "client": {
+                "base_url": "https://pokeapi.co/api/v2/",
+                "default_paginator": "json_links",
+            },
+            "endpoints": {
+                "pokemon": {
+                    "params": {
+                        "limit": 1000,  # Default page size is 20
+                    },
+                },
+                "berry": {
+                    "params": {
+                        "limit": 1000,
+                    },
+                },
+                "location": {
+                    "params": {
+                        "limit": 1000,
+                    },
                 },
             },
-            "berry": {
-                "params": {
-                    "limit": 1000,
-                },
-            },
-            "location": {
-                "params": {
-                    "limit": 1000,
-                },
-            },
-        },
-    })
+        }
+    )
 
     load_info = pipeline.run(pokemon_source)
     print(load_info)
